renamed downloadAlbum -> downloadPlaylist
